requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
gunicorn
uvicorn
selenium
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic